style: unify watering UI styles between table and cards
