+++ frontend/src/components/trees/watering/WateringCard.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-	import { formatDate, waterEmoji } from '$lib/utils/watering';
+	import { formatDate } from '$lib/utils/formatDate';
+	import { waterEmoji } from '$lib/waterings';
+
 	import type { WateringWithTree } from '$types/watering';
 
 	export let watering: WateringWithTree;
@@ -7,31 +9,31 @@
 	export let currentUserId: string | null;
 </script>
 
-<div class="bg-white rounded-lg shadow-sm p-4 text-sm">
+<div class="bg-white border-2 border-[#7C98B2] rounded-xl shadow-sm p-4 text-sm">
 	{#if mode === 'user' && watering.tree?.uuid}
-		<div class="flex justify-center mb-4">
+		<div class="flex justify-left mb-4">
 			<slot name="treeButton" {watering} />
 		</div>
 	{/if}
 
-	<div class="flex justify-between text-gray-500 mb-2">
-		<span class="font-semibold">Datum</span>
+	<div class="flex items-center justify-between text-gray-600 mb-2">
+		<span class="text-xs tracking-wide font-medium text-gray-500">🗓️ Datum</span>
 		<span class="text-black">{formatDate(watering.watered_at)}</span>
 	</div>
 
-	<div class="flex justify-between text-gray-500 mb-2">
-		<span class="font-semibold">Liter</span>
+	<div class="flex items-center justify-between text-gray-600 mb-2">
+		<span class="text-xs tracking-wide font-medium text-gray-500">🚰 Liter</span>
 		<span class="text-black">{watering.amount_liters}</span>
 	</div>
 
-	<div class="flex justify-between text-gray-500 mb-2">
-		<span class="font-semibold">Gießkraft</span>
+	<div class="flex items-center justify-between text-gray-600 mb-2">
+		<span class="text-xs tracking-wide font-medium text-gray-500">🚿 Gießkraft</span>
 		<span class="text-black">{waterEmoji(watering.amount_liters)}</span>
 	</div>
 
 	{#if mode === 'tree'}
-		<div class="flex justify-between text-gray-500 mb-2">
-			<span class="font-semibold">Durch</span>
+		<div class="flex items-center justify-between text-gray-600 mb-2">
+			<span class="text-xs tracking-wide font-medium text-gray-500">👤 Durch</span>
 			<span class="text-black">
 				<em>{currentUserId && watering.user_uuid === currentUserId ? 'Du' : 'anonym'}</em>
 			</span>
@@ -39,7 +41,7 @@
 	{/if}
 
 	{#if currentUserId && watering.user_uuid === currentUserId}
-		<div class="mt-3 flex justify-center">
+		<div class="mt-2 flex justify-end">
 			<slot name="deleteButton" {watering} />
 		</div>
 	{/if}
