+++ frontend/src/lib/waterings/index.ts
@@ -1,0 +1,1 @@
+export * from './visual';
