+++ frontend/src/lib/waterings/visual.ts
@@ -1,0 +1,7 @@
+// Emoji-Darstellung der Wassermenge (visuelle Gießkraft)
+export function waterEmoji(amount: number): string {
+	if (amount >= 40) return '💧💧💧';
+	if (amount >= 20) return '💧💧';
+	if (amount >= 5) return '💧';
+	return '🌱';
+}
